third_party/WebKit/LayoutTests/http/tests/websocket/nocache.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/get-request-header.js"></script>
<script>
description('Test whether WebSocket request contains no-cache headers');

window.jsTestIsAsync = true;

var pragma;
var cache_control;

connectAndGetRequestHeaders().then(function(headers) {
    pragma = headers['pragma'];
    cache_control = headers['cache-control'];
    shouldBeEqualToString('pragma', 'no-cache');
    shouldBeEqualToString('cache_control', 'no-cache');
    finishJSTest();
}, finishAsFailed);
</script>
