third_party/blink/web_tests/http/tests/websocket/nocache-expected.txt

Test whether WebSocket request contains no-cache headers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pragma is "no-cache"
PASS cache_control is "no-cache"
PASS successfullyParsed is true

TEST COMPLETE
